Restructure solution, update YAML
