+++ LeetCode.Submissions/Easy/TwoSumService.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace LeetCode.Easy.Submissions
+namespace LeetCode.Submissions.Easy
 {
     /// <summary>
     /// Given an array of integers nums and an integer target, return indices of the two numbers such that they add up to target.
