+++ LeetCode.Submissions/Medium/ReverseInteger.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace LeetCode.Easy.Submissions
+namespace LeetCode.Submissions.Medium
 {
     public static class ReverseInteger
     {
