--- azure-pipelines.yml
@@ -1,30 +1,0 @@
-# Starter pipeline
-# Start with a minimal pipeline that you can customize to build and deploy your code.
-# Add steps that build, run tests, deploy, and more:
-# https://aka.ms/yaml
-
-name: Build and Test
-
-trigger: [main]
-pr: [ main ]
-
-
-pool:
-  vmImage: ubuntu-latest
-
-variables:
-  buildConfiguration: 'Release'
-
-
-steps:
-- task: UseDotNet@2
-  displayName: 'Install .NET 6'
-  inputs:
-    version: '6.x'
-    includePreviewVersions: false # Required for preview versions
-
-- task: DotNetCoreCLI@2
-  displayName: 'Build Project - Configuration $(buildConfiguration)'
-  inputs:
-    command: 'build'
-    arguments: '--configuration $(buildConfiguration)'
